+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+# [Unreleased]
+### Changed
+- Remove HEAD query and cache from get_full_res_image_url method.
+
 # [2.1.0] - 2020-12-17
 ### Added
 - Analytics usage may now be switched on/off with an environment variable